Add files via upload
This is gui interface in java for generating tillers (Romanian: motosape), filtering them, saving the generated list on disk and then loading the last saved list back into our text area. And then we are able to filter them all over again. The gui is a work in progress, and I am going to add a few more functionalitiess in the subsequent months.